+++ app/tags/controller.js
@@ -3,6 +3,7 @@ const config = require('../config');
 
 async function index(req, res, next) {
     try {
+        
         let {
             limit = 10,
                 skip = 0
@@ -21,6 +22,15 @@ async function index(req, res, next) {
 
 async function store(req, res, next) {
     try {
+        let policy = policyFor(req.user)
+
+        if (!policy.can('create', 'Tag')) {
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk membuat Tag`
+            });
+        }
+
         let payload = req.body
 
         let tag = new Tag(payload)
@@ -42,6 +52,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
     try {
+        let policy = policyFor(req.user)
+
+        if (!policy.can('update', 'Tag')) {
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk mengupdate Tag`
+            });
+        }
+
         let payload = req.body;
 
         let tag = await Tag.findOneAndUpdate({
@@ -66,6 +85,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
     try {
+        let policy = policyFor(req.user)
+
+        if (!policy.can('delete', 'Category')) {
+            return res.json({
+                error: 1,
+                message: `Anda tidak memiliki akses untuk menghapus Tag`
+            });
+        }
+
         let tag = await Tag.findOneAndDelete({
             _id: req.params.id
         });
